% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTau.R
\name{calcTau}
\alias{calcTau}
\title{Implements tissue specificity algorithm}
\usage{
calcTau(x)
}
\arguments{
\item{x}{quantile normalised data frame i.e. output from quantNorm()}
}
\value{
Returns a data frame of tau expression fractions indicating the specificty 
of a gene for a tissue between 0 (not specific) and 1 (absolutely specific)
}
\description{
This function first places all quantile normalised values into decile bins. 
For each gene, it then uses the bin profile (row of bin values) to calculate 
a single tau value between 0 (non-specific) and 1 (absolutely specific). 
Finally, for each gene it uses the tau value + quantile normalised counts 
to identify a tau expression fraction between 0 (non-specific) and 1 
(absolutely specific) which indicates the specificity of a single gene for 
a single tissue
}
\examples{
head(qnExp)
tauExp <- calcTau(qnExp)
head(tauExp)
}
