% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Checks the input data frame for incompatible values
values, and infinite values}
\usage{
checkInput(x)
}
\arguments{
\item{x}{data frame of mean expression (cells) of genes (rows) in 
tissues (columns)}
}
\value{
Returns errors when incompatible values are found. Also returns a 
printout of the number of errors in each column. Nothing is returned If the 
dataframe is compatible with the package
}
\description{
The input dataframe contains expression data so each cell should contain only 
non-negative, non-infinite, numerical characters (>=0). This function 
identifies if the dataframe contains character strings, NAs, negative values 
or infinite values that must first be removed by the user.
}
\examples{
## check example data
checkInput(meanExp)

## check data containing character strings
#df <- data.frame('col1' = c(1,'foo',3), 'col2' = c(4,5,'bar'))
#checkInput(df)

## check data containing other incompatible values
#df <- data.frame('col1' = c(1,2,Inf), 'col2' = c(-4,-Inf,6))
#checkInput(df)
}
