% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getControls.R
\name{getControls}
\alias{getControls}
\title{Get a set of genes that can be used to differentiate between 2 tissues}
\usage{
getControls(x, y)
}
\arguments{
\item{x}{primary tissue, output from getTissue()}

\item{y}{secondary tissue, output from getTissue()}
}
\value{
Returns a list object containing 1 dataframe of genes for each of the 2 
tissues.
}
\description{
This function selects 2 set of genes, 1 for each tissue. Each set includes 
genes that are highly specific and expressed in one tissue with <0.1 
quantile normalised expression in the other. This is useful for 
differentiating between tissue types. For example, for selecting positive & 
negative controls for qPCR to determine if a tissue sample has been 
contaminated.
}
\examples{
# Usage
controls <- getControls(tissueA, tissueB)
# To confirm the difference in expression between tissues
x <- rownames(head(controls$tissueA, n = 1))
qnExp[x, ]
tissueA[x, ]
tissueB[x, ]
}
