% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist.R
\name{getDist}
\alias{getDist}
\title{Get tissue specific distribution}
\usage{
getDist(x, y)
}
\arguments{
\item{x}{ensembl annotated tau expression fraction file 
i.e. output from getMart()}

\item{y}{threshold tau value, range 0-1}
}
\value{
Returns the number of each type of tissue specific gene that 
meet the user defined specificity threshold
}
\description{
This function counts the total number of tissue specific protein coding, 
lincRNA, and miRNA genes for each tissue.
}
\examples{
# Get the number of absolutely specific genes (ASGs, tau = 1) in each tissue
head(tauAnno)
getDist(tauAnno, 1) 
# Get the number of highly specific genes (HSGs, tau = 0.85) in each tissue
getDist(tauAnno, 0.85) 
}
