% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMart.R
\name{getMart}
\alias{getMart}
\title{Get ensembl annotations}
\usage{
getMart(x, y, z)
}
\arguments{
\item{x}{species: 'human', 'macaque', or 'mouse'}

\item{y}{ensembl archive version: 79..90}

\item{z}{output from calcTau()}
}
\value{
Returns a dataframe containing tau values, tau expression fractions, 
ensembl gene ids, names, and types in the selected species from the 
selected ensembl archive version.
}
\description{
Connects to ensembl archive to get id, name and type for every gene
}
\examples{
head(tauExp)
tauAnno <- getMart(x = 'mouse', y = 79, z = tauExp)
head(tauAnno)
}
