% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissue.R
\name{getTissue}
\alias{getTissue}
\title{Get expression and specificity of each gene in a single tissue}
\usage{
getTissue(x, y, z)
}
\arguments{
\item{x}{tissue of interest (case sensitive)}

\item{y}{output from quantNorm()}

\item{z}{output from getMart()}
}
\value{
Returns a data frame of ensembl annotations, normalised expression 
and specificity values
}
\description{
This function takes a user defined tissue and extracts the ensembl 
annotations, normalised expression values, tau values, and tau expression
fractions of all genes in that tissue.
}
\examples{
x <- head(colnames(meanExp), n = 1)
x <- getTissue(x, qnExp, tauAnno)
head(x)
y <- rownames(head(x, n = 1))
qnExp[y,]
tauAnno[y,]
}
