% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2Exp.R
\docType{data}
\name{log2Exp}
\alias{log2Exp}
\title{Example log normalised data set output of log2Trans()}
\format{Dataframe of ensembl_gene_ids, and their log2(rpkm) in each tissue}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
log2Exp
}
\description{
This example data set is provided to allow the user to run the examples 
in the documentation of each package function
}
\keyword{datasets}
