% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2Tran.R
\name{log2Tran}
\alias{log2Tran}
\title{Log Transforms a data frame}
\usage{
log2Tran(x)
}
\arguments{
\item{x}{data frame of mean expression (cells) of genes (rows) in 
tissues (columns)}
}
\value{
Returns a data frame of log2 normalised counts, with non-expressed genes 
(rows) removed
}
\description{
Adds a pseudocount (1e-10) to remove zeros, then log2 transforms each 
column independently. This normalises the data frame within each column 
only (not across columns - see tispec::QuantNorm). After log2 
transformation, it resets all negative values to zero, thus setting a 
noise threshold of log2(1). Finally it removes all rows where every column 
is 0 (i.e. below threshold, removing genes not expressed in any tissue)
}
\examples{
head(meanExp)
log2Exp <- log2Tran(meanExp)
head(log2Exp)
}
