% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanExp.R
\docType{data}
\name{meanExp}
\alias{meanExp}
\title{Example input data set}
\format{Dataframe of ensembl_gene_ids, their rpkm in each tissue}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
meanExp
}
\description{
This example data set is provided to allow the user to run the examples 
in the documentation of each package function
}
\keyword{datasets}
