% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimum.R
\docType{data}
\name{optimum}
\alias{optimum}
\title{Example output of getOptimum()}
\format{A list containing 2 objects: dataframe and barplot.}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
optimum
}
\description{
This example data set is provided to allow the user to run the examples in 
the documentation of each package function
}
\keyword{datasets}
