% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorr.R
\name{plotCorr}
\alias{plotCorr}
\title{Plot correlation between expression and specificity}
\usage{
plotCorr(x)
}
\arguments{
\item{x}{output from getTissue()}
}
\value{
Returns a list object containing 2 ggplot objects: global, local. Global is 
the result of plotting QN expression v the overall specificity of each gene 
for ANY tissue (tau). Local is the result of plotting QN expression v the 
specificity of each gene for the tissue being plotted (tau expression 
fraction).

Yellow: tau expression fraction >= 0.85. 

Orange: tau expression fraction = 1. 

Red: generalised additive model trend line

Swatch: 0.99 confidence interval

r: correlation coefficient
}
\description{
This function plots quantile normalised expression values versus
tissue specificity values of all genes in all tissues to visualise 
correlation
}
\examples{
corrPlots <- plotCorr(tissueA)
#corrPlots$global
#corrPlots$local
}
