% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot tissue specificity distribution}
\usage{
plotDensity(x)
}
\arguments{
\item{x}{data frame of annotated tissue specificity (tau) values 
i.e. output from getMart()}
}
\value{
Returns a tissue specificity density plot
}
\description{
This function plots the tau value of every gene to see which tau values 
occur most often
}
\examples{
head(tauAnno)
plotDensity(tauAnno)
}
