% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Plot number of tissue specific genes per tissue}
\usage{
plotDist(x)
}
\arguments{
\item{x}{ensembl annotated tau expression fraction file 
i.e. output from getMart()}
}
\value{
Returns a barplot of tissue specific gene counts per tissue
}
\description{
This function plots the number of absolutely specific genes (ASGs, tau = 1)
alongside the number of highly specific genes (HSGs, tau >=0.85) in 
each tissue
}
\examples{
head(tauAnno)
plotDist(tauAnno)
}
