% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot specificity of a user defined gene for each tissue}
\usage{
plotGene(x, y)
}
\arguments{
\item{x}{ensembl annotated tau expression fraction file 
i.e. output from getMart()}

\item{y}{user define gene, case dependent, to see gene available names 
view output of getMart()}
}
\value{
Returns a bar plot of tau espression fraction in each tissue
}
\description{
This function takes a user defined gene (case sensitive) and plots the
tau expression fraction of that gene in each tissue. This allows the user 
to visualize the tissue/tissues in which the genes has relatively 
enriched/depleted expression
}
\examples{
x <- head(tauAnno$external_gene_name, n = 1)
plotGene(tauAnno, x)
y <- rownames(subset(tauAnno, tauAnno$external_gene_name == x))
meanExp[y,]
qnExp[y,]
tauAnno[y,]
}
