% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTop.R
\name{plotTop}
\alias{plotTop}
\title{Plot a set of highly specific, highly expressed genes in for 
tissue of interest}
\usage{
plotTop(x, y, z)
}
\arguments{
\item{x}{output from getMart()}

\item{y}{output from getTissue()}

\item{z}{number of genes}
}
\value{
Returns a barplot showing all tissues in which the set of genes are 
expressed and the specificity of those genes for those tissues. 
Red dotted line: 0.85 tau expression fraction cut off for high tissue 
specificity
}
\description{
This function extracts all genes highly specific to the tissue of 
interest (tauExpFrac >=0.85) and ranks them by expression
It then selects z number of top expressed genes and plots tau expression 
fraction of those genes in all tissues they are expressed in
}
\examples{
plotTop(tauAnno, tissueA, 5)
}
