% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnExp.R
\docType{data}
\name{qnExp}
\alias{qnExp}
\title{Example quantile normalised data set output of quantNorm()}
\format{Dataframe of ensembl_gene_ids, and their quantile normalised
expression across all tissues}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
qnExp
}
\description{
This example data set is provided to allow the user to run the examples 
in the documentation of each package function
}
\keyword{datasets}
