% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantNorm.R
\name{quantNorm}
\alias{quantNorm}
\title{Quantile normalise a data frame}
\usage{
quantNorm(x)
}
\arguments{
\item{x}{log transformed data frame of mean expression (cells) of 
genes (rows) in tissues (columns)}
}
\value{
Returns a data frame of equal dimensions with quantile normalised 
counts across all columns
}
\description{
This function quantile normalises the entire data frame of log normalised 
counts, allowing comparisons across tissues. First, the row and column 
names are collected and all zeros are changed to NA. After converting to a 
matrix, the data frame is quantile normalised. Finally, all NAs are 
reverted back to zero, the matrix is converted back to a data frame and 
row and column names are re-attached
}
\examples{
head(log2Exp)
qnExp <- quantNorm(log2Exp)
head(qnExp)
}
