% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauExp.R
\docType{data}
\name{tauExp}
\alias{tauExp}
\title{Example tau + expression fraction output of calcTau()}
\format{Dataframe of ensembl_gene_ids, their global tau values, 
and their tau expression fractions in each tissue. Tau expression fraction 
is a measure of specificity of a single gene for a single tissue.}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
tauExp
}
\description{
This example data set is provided to allow the user to run the examples 
in the documentation of each package function
}
\keyword{datasets}
