% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueA.R
\docType{data}
\name{tissueA}
\alias{tissueA}
\title{Example tissueA output of getTissue()}
\format{A dataframe of ensembl_gene_ids and their names, biotypes, 
quantile normalized expression in tissueA, global tau specificity values and 
tau expression fraction for tissueA}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
tissueA
}
\description{
This example data set is provided to allow the user to run the examples in 
the documentation of each package function
}
\keyword{datasets}
