% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueB.R
\docType{data}
\name{tissueB}
\alias{tissueB}
\title{Example tissueB output of GetTissue()}
\format{A dataframe of ensembl_gene_ids and their names, biotypes, 
quantile normalized expression in tissueB, global tau specificity values and 
tau expression fraction for tissueB}
\source{
The data frame has been randomly generated, no biological 
conclusions can be made from this data set.
}
\usage{
tissueB
}
\description{
This example data set is provided to allow the user to run the examples in 
the documentation of each package function
}
\keyword{datasets}
